docs: Container usage
Replace doc about npm install with this container, assuming that is more familiar to end-users. This also later allows to bundle additional tools into the container image, without assuming that they are locally available.

Also adds a short new Developers section, but folks who want to not just use but hack on this project are likely already familiar with what's what anyway.